requirements.txt
Flask
flask-cors
transformers
torch
scikit-learn
nltk
pillow
opencv-python